refactor/traverse: introduce unit tests for module traversal
