+++ demo/src/main.rs
@@ -1,3 +1,4 @@
+//! Just a lil somethin somethin
 mod lib;
 
 pub mod nested;
